Changes to move the tests to inside tox
