+++ .github/workflows/cd.yml
@@ -1,0 +1,91 @@
+
+name: CD
+
+on:
+  push:
+    tags:
+      - "v[0-9]+.[0-9]+.[0-9]+*"
+
+env:
+  FORCE_COLOR: 1
+
+jobs:
+  pre-commit:
+    runs-on: ubuntu-latest
+    timeout-minutes: 90
+    strategy:
+      matrix:
+        python: ["3.8"]
+    steps:
+      - uses: actions/checkout@v3.3.0
+      - name: "Install dependencies"
+        run: sudo apt-get -y install graphviz
+      - name: Cache python dependencies
+        id: cache-pip
+        uses: actions/cache@v3.2.4
+        with:
+          path: ~/.cache/pip
+          key: pip-${{ matrix.python }}-tests-${{ hashFiles('**/setup.json') }}
+          restore-keys: pip-${{ matrix.python }}-tests-
+      - name: Set up Python
+        uses: actions/setup-python@v4.5.0
+        with:
+          python-version: ${{ matrix.python }}
+      - name: Make sure virtualevn>20 is installed, which will yield newer pip and possibility to pin pip version.
+        run: pip install virtualenv>20
+      - name: Install Tox
+        run: pip install tox
+      - name: Run pre-commit in Tox
+        run: tox -e pre-commit
+  test:
+    runs-on: "ubuntu-latest"
+    strategy:
+      matrix:
+        python: ["3.8", "3.9", "3.10"]
+    steps:
+    - uses: actions/checkout@v3.3.0
+    - name: "Install dependencies"
+      run: sudo apt-get -y install graphviz
+      - uses: actions/checkout@v3.3.0
+      - name: Cache python dependencies
+        id: cache-pip
+        uses: actions/cache@v3.2.4
+        with:
+          path: ~/.cache/pip
+          key: pip-${{ matrix.python }}-tests-${{ hashFiles('**/setup.json') }}
+          restore-keys: pip-${{ matrix.python }}-tests-
+    - name: Set up Python
+      uses: actions/setup-python@v4.5.0
+      with:
+        python-version: ${{ matrix.python }}
+    - name: Make sure virtualevn>20 is installed, which will yield newer pip and possibility to pin pip version.
+      run: pip install virtualenv>20
+    - name: Install Tox
+      run: pip install tox
+    - name: "Test with tox"
+      run: |
+        tox -e py${{ matrix.python }} -- tests/ --cov=./sqalchemy_schemadisplay --cov-append --cov-report=xml --cov-report=term-missing
+
+  release:
+    if: "github.event_name == 'push' && startsWith(github.ref, 'refs/tags')"
+    runs-on: "ubuntu-latest"
+    needs: "test"
+    steps:
+    - uses: "actions/checkout@v3"
+
+    - name: "Install dependencies"
+      run: |
+        python -m pip install --upgrade pip
+        pip install build
+
+    - name: "Build"
+      run: |
+        python -m build
+        git status --ignored
+
+    - name: "Publish"
+      uses: "pypa/gh-action-pypi-publish@release/v1"
+      with:
+        user: "__token__"
+        password: "${{ secrets.TEST_PYPI_API_TOKEN }}"
+        repository_url: "https://test.pypi.org/legacy/"
