--- tox.ini
@@ -1,31 +1,0 @@
-[tox]
-
-[testenv]
-commands = py.test --cov-report html --cov-report term --cov sqlalchemy_schemadisplay {posargs}
-deps =
-    pytest
-    pytest-cov
-
-[testenv:sqla06-py27]
-deps =
-    {[testenv]deps}
-    sqlalchemy==0.6.*
-
-[testenv:sqla07-py27]
-deps =
-    {[testenv]deps}
-    sqlalchemy==0.7.*
-
-[testenv:sqla08-py27]
-deps =
-    {[testenv]deps}
-    sqlalchemy==0.8.*
-
-[testenv:sqla09-py27]
-deps =
-    {[testenv]deps}
-    sqlalchemy==0.9.*
-
-[testenv:sqlalchemy-py27]
-
-[testenv:sqlalchemy-py3]
